Move pylxd.utils to pylxd.deprecated.utils
