+++ pylxd/deprecated/utils.py
